wine-ml-app/backend/requirements.txt
fastapi
uvicorn
mlflow
scikit-learn
imbalanced-learn
pandas
numpy
shap
xgboost
requests
python-multipart
python-dotenv
optuna
supabase
